cart

+++ src/controllers/frontend/cart-order-controller.ts
@@ -1,5 +1,5 @@
 import 'module-alias/register';
-import { Request, Response } from 'express';
+import { Request, Response, query } from 'express';
 
 import { capitalizeWords, formatZodError, handleFileUpload, slugify } from '../../utils/helpers';
 import { cartProductSchema, cartSchema } from '../../utils/schemas/frontend/guest/cart-schema';
@@ -22,6 +22,7 @@ import { addToWishlistSchema } from '../../utils/schemas/frontend/auth/wishlist-
 import CustomerWishlistCountryService from '../../services/frontend/auth/customer-wishlist-servicel'
 import WebsiteSetupModel from '../../model/admin/setup/website-setup-model';
 import { blockReferences, websiteSetup } from '../../constants/website-setup';
+import { QueryParams } from '../../utils/types/common';
 
 const controller = new BaseController();
 
@@ -103,6 +104,7 @@ class CartController extends BaseController {
                                     giftWrapAmount: item.giftWrapAmount,
                                     quantity: item.quantity,
                                     variantId: productVariantData._id,
+                                    productId: productVariantData.productId,
 
                                 };
                             } else {
@@ -119,37 +121,69 @@ class CartController extends BaseController {
                         var updateCart
 
                         for (let data of combinedData) {
-
-                            updateCart = await CartService.updateCartProductByCart({
+                            console.log("khkhjkh", data);
+                            const cartProduct: any = await CartOrderProductsModel.findOne({
                                 $and: [
                                     { cartId: data.cartId },
                                     { variantId: data.variantId }
                                 ]
-                            }, data)
-
-
-                        }
-                        if (updateCart) {
-
-                            const deletedData = await CartService.destroy(guestUserCart._id);
-                            const updateData = await CartService.update(customerCart._id, { guestUserId: null });
-
-                            const deletedProductData = await CartService.destroyCartProduct1({ cartId: guestUserCart._id });
-
-                            const cart = await CartService.findCartPopulate({
-                                $and: [
-                                    { customerId: customer },
-                                    { countryId: country }
-                                ]
                             })
+                            console.log("cartProduct", cartProduct);
 
-                            return controller.sendSuccessResponse(res, {
-                                requestedData: {
-                                    ...cart
+                            if (cartProduct) {
+                                updateCart = await CartService.updateCartProductByCart({
+                                    $and: [
+                                        { cartId: data.cartId },
+                                        { variantId: data.variantId }
+                                    ]
                                 },
-                                message: 'Your cart is ready!'
-                            });
+                                    {
+                                        cartId: data.cartId,
+                                        slug: data.slug,
+                                        giftWrapAmount: data.giftWrapAmount,
+                                        quantity: data.quantity,
+                                        variantId: data.variantId,
+                                        productId: data.productId
+                                    })
+                                console.log("dddddddddddd", updateCart);
+
+                                if (updateCart) {
+
+                                    const deletedData = await CartService.destroy(guestUserCart._id);
+                                    const updateData = await CartService.update(customerCart._id, { guestUserId: null });
+
+                                    const deletedProductData = await CartService.destroyCartProduct1({ cartId: guestUserCart._id });
+
+                                    const cart: any = await CartService.findCartPopulate({
+                                        query: {
+                                            $and: [
+                                                { customerId: customer },
+                                                { countryId: country }
+                                            ]
+                                        },
+                                        hostName: req.get('origin'),
+                                    })
+
+                                    // const cart = await CartService.findCartPopulate({
+                                    //     $and: [
+                                    //         { customerId: customer },
+                                    //         { countryId: country }
+                                    //     ]
+                                    // })
+
+                                    return controller.sendSuccessResponse(res, {
+                                        requestedData: {
+                                            ...cart
+                                        },
+                                        message: 'Your cart is ready!'
+                                    });
+                                }
+                            }
+
+
+
                         }
+
                     }
                 } else
 
@@ -168,6 +202,8 @@ class CartController extends BaseController {
                         let totalAmountOfProduct = 0
                         let totalDiscountAmountOfProduct = 0
                         let quantityProduct = 0
+                        totalAmountOfProduct = productVariantData?.price * quantity
+                        totalDiscountAmountOfProduct = productVariantData.discountPrice * quantity
 
                         if (existingCart) {
 
@@ -181,11 +217,11 @@ class CartController extends BaseController {
 
 
                             if (existingCart.totalProductAmount) {
-                                totalAmountOfProduct = existingCart.totalProductAmount + (productVariantData.price * quantity)
-                                totalDiscountAmountOfProduct = existingCart.totalDiscountAmount + (productVariantData.price * quantity)
+                                totalAmountOfProduct = (productVariantData.price * existingCartProduct.quantity)
+                                totalDiscountAmountOfProduct = (productVariantData.discountPrice * existingCartProduct.quantity)
                             } else {
                                 totalAmountOfProduct = productVariantData?.price * quantity
-                                totalDiscountAmountOfProduct = productVariantData.price * quantity
+                                totalDiscountAmountOfProduct = productVariantData.discountPrice * quantity
 
                             }
 
@@ -201,7 +237,7 @@ class CartController extends BaseController {
                                     const deletedData = await CartService.destroyCartProduct(existingCartProduct._id);
                                     if (deletedData) {
 
-                                        const cart = await CartService.findCartPopulate({ _id: existingCartProduct.cartId })
+                                        const cart = await CartService.findCartPopulate({ _id: existingCartProduct.cartId, hostName: req.get('origin') })
 
                                         return controller.sendSuccessResponse(res, {
                                             requestedData: {
@@ -235,8 +271,10 @@ class CartController extends BaseController {
                             }
                         }
 
+
                         const shippingAmount: any = await WebsiteSetupModel.findOne({ blockReference: blockReferences.shipmentSettings })
                         // const codAmount: any = await WebsiteSetupModel.findOne({ blockReference: blockReferences.defualtSettings })
+                        // console.log(productVariantData, "........1234d.....", codAmount);
 
                         const cartOrderData = {
                             customerId: customer,
@@ -259,8 +297,9 @@ class CartController extends BaseController {
                             totalCouponAmount,
                             totalWalletAmount,
                             codAmount,
+                            // codAmount: Number(codAmount.blockValues.codCharge),
                             totalTaxAmount,
-                            totalAmount: totalAmountOfProduct - totalDiscountAmountOfProduct,
+                            totalAmount: totalAmountOfProduct + Number(shippingAmount.blockValues.shippingCharge),
                         };
 
                         if (existingCart) {
@@ -279,6 +318,7 @@ class CartController extends BaseController {
                                     cartId: newCartOrder._id,
                                     customerId: customer,
                                     variantId: productVariantData._id,
+                                    productId: productVariantData.productId,
                                     quantity: quantityProduct,
                                     slug: productVariantData.slug,
                                     orderStatus,
@@ -306,6 +346,7 @@ class CartController extends BaseController {
                                     cartId: newCartOrder._id,
                                     customerId: customer,
                                     variantId: productVariantData._id,
+                                    productId: productVariantData.productId,
                                     quantity: quantityProduct ? 0 : quantity,
                                     slug: productVariantData.slug,
                                     orderStatus,
@@ -439,10 +480,14 @@ class CartController extends BaseController {
             const { variantId, slug } = req.body;
 
             const cart: any = await CartService.findCartPopulate({
-                $or: [
-                    { $and: [{ customerId: customer }, { countryId: country }] },
-                    { $and: [{ guestUserId: guestUser }, { countryId: country }] }
-                ]
+                query: {
+                    $or: [
+                        { $and: [{ customerId: customer }, { countryId: country }] },
+                        { $and: [{ guestUserId: guestUser }, { countryId: country }] }
+                    ]
+                },
+                hostName: req.get('origin'),
+
             });
 
             const existingCartProduct: any = await CartService.findCartProduct({
@@ -455,7 +500,7 @@ class CartController extends BaseController {
             const giftWrapAmount: any = await WebsiteSetupModel.findOne({ blockReference: blockReferences.enableFeatures })
             var giftWrapCharge
             if (giftWrapAmount.blockValues.enableGiftWrap == true) {
-                giftWrapCharge = giftWrapAmount.blockValues.giftWrapCharge
+                giftWrapCharge = Number(giftWrapAmount.blockValues.giftWrapCharge)
             }
 
 
@@ -467,10 +512,10 @@ class CartController extends BaseController {
                 ]
             }, { giftWrapAmount: giftWrapCharge })
 
-            const cartUpdate: any = await CartService.update(cart._id, { totalGiftWrapAmount: giftWrapCharge });
+            const cartUpdate: any = await CartService.update(cart._id, { totalGiftWrapAmount: (cart.totalGiftWrapAmount + giftWrapCharge), totalAmount: (cart.totalAmount + giftWrapCharge) });
 
             return controller.sendSuccessResponse(res, {
-                requestedData: cartUpdate,
+                requestedData: cart,
                 message: 'gift wrap added successfully!'
             });
 
@@ -483,22 +528,38 @@ class CartController extends BaseController {
 
     async findUserCart(req: Request, res: Response): Promise<void> {
         try {
-
+            const { page_size = 1, limit = 20, sortby = '', sortorder = '' } = req.query as QueryParams;
             const customer = res.locals.user;
             const guestUser = res.locals.uuid;
             let country = await CommonService.findOneCountrySubDomainWithId(req.get('origin'));
+            // let query = {
+            //     $or: [
+            //         { $and: [{ customerId: customer }, { countryId: country }] },
+            //         { $and: [{ guestUserId: guestUser }, { countryId: country }] }
+            //     ]
+            // }
 
+            let query: any = { _id: { $exists: true } };
+            const userData = await res.locals.user;
+
+            // query.status = '1';
+            query.customerId = userData._id;
+            console.log(query);
+
+            const sort: any = {};
+            if (sortby && sortorder) {
+                sort[sortby] = sortorder === 'desc' ? -1 : 1;
+            }
             const cart: any = await CartService.findCartPopulate({
-                $or: [
-                    { $and: [{ customerId: customer }, { countryId: country }] },
-                    { $and: [{ guestUserId: guestUser }, { countryId: country }] }
-                ]
+                page: parseInt(page_size as string),
+                limit: parseInt(limit as string),
+                query,
+                hostName: req.get('origin'),
+                sort
             });
 
             return controller.sendSuccessResponse(res, {
-                requestedData: {
-                    ...cart
-                },
+                requestedData: cart,
                 message: 'Your cart is ready!'
             });
         } catch (error: any) {
